refactor: Move Connection to libs
